MotionBack-Dev/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.101.0
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.27.1
